move clock and related structs into new crate
