+++ Cargo.toml
@@ -2,6 +2,7 @@
 resolver = "2"
 members = [
     "crates/alator",
+    "crates/alator-clock",
     "crates/alator-exchange",
     "integrations/snake",
 ]
