+++ crates/alator-clock/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "alator-clock"
+version = "0.1.0"
+edition = "2021"
+
+
+[dependencies]
+
+time = { version = "0.3.17", features = ["macros", "parsing"] }
