+++ crates/alator/Cargo.toml
@@ -10,7 +10,6 @@ readme = "README.md"
 
 [dependencies]
 itertools = "0.11.0"
-time = { version = "0.3.17", features = ["macros", "parsing"] }
 rand = "0.8.4"
 rand_distr = "0.4.1"
 log = "0.4.17" 
@@ -19,6 +18,7 @@ pyo3 = { version = "0.20.0", optional = true }
 async-trait = "0.1.73"
 tokio = { version = "1.32.0", features = ["full"] }
 futures = "0.3.28"
+alator-clock = { path = "../alator-clock" }
 
 [dev-dependencies]
 reqwest = { version = "0.11.11", features=["blocking"] }
