+++ crates/alator/src/broker/record.rs
@@ -1,7 +1,8 @@
 use itertools::Itertools;
 
 use super::{BrokerRecordedEvent, DividendPayment, Trade, TradeType};
-use crate::types::{CashValue, DateTime, PortfolioQty, Price};
+use crate::types::{CashValue, PortfolioQty, Price};
+use alator_clock::DateTime;
 
 //Records events generated by brokers. Used for internal calculations but is public for tax
 //calculations.
