+++ crates/alator/src/exchange/implement/mod.rs
@@ -47,7 +47,7 @@
 //! for communications that are relevant to individual brokers, for example completed trades.
 
 #[allow(unused)]
-use crate::clock::Clock;
+use alator_clock::Clock;
 #[allow(unused)]
 use crate::input::PriceSource;
 #[allow(unused)]
