+++ crates/alator/src/exchange/implement/multi/builder.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use crate::clock::Clock;
+use alator_clock::Clock;
 use crate::exchange::implement::multi::ConcurrentExchange;
 use crate::input::{PriceSource, Quotable};
