+++ crates/alator/src/exchange/implement/single/builder.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use crate::clock::Clock;
+use alator_clock::Clock;
 use crate::exchange::implement::single::SingleExchange;
 use crate::input::{PriceSource, Quotable};
