+++ crates/alator/src/perf/mod.rs
@@ -1,7 +1,8 @@
 //! Generates performance stats for backtest
 
-use crate::types::{Frequency, StrategySnapshot};
+use crate::types::StrategySnapshot;
 use itertools::Itertools;
+use alator_clock::Frequency;
 
 /// Output for single backtest run.
 #[derive(Clone, Debug)]
@@ -256,7 +257,7 @@ impl PerformanceCalculator {
 mod tests {
     use crate::broker::implement::single::{SingleBroker, SingleBrokerBuilder};
     use crate::broker::{BrokerCost, Dividend, Quote};
-    use crate::clock::{Clock, ClockBuilder};
+    use alator_clock::{Clock, ClockBuilder};
     use crate::exchange::implement::single::SingleExchangeBuilder;
     use crate::input::{DefaultCorporateEventsSource, DefaultPriceSource};
     use crate::perf::StrategySnapshot;
