+++ crates/alator/src/strategy/implement/staticweight/builder.rs
@@ -1,6 +1,6 @@
 use crate::broker::implement::multi::ConcurrentBroker;
 use crate::broker::implement::single::SingleBroker;
-use crate::clock::Clock;
+use alator_clock::Clock;
 use crate::input::{CorporateEventsSource, Dividendable, PriceSource, Quotable};
 use crate::types::PortfolioAllocation;
