+++ integrations/snake/Cargo.toml
@@ -11,5 +11,6 @@ crate-type = ["cdylib"]
 [dependencies]
 pyo3 = { version = "0.20.0", features = ["extension-module"]}
 alator = { path = "../../crates/alator", features = ["python"] }
+alator-clock = { path = "../../crates/alator-clock" }
 rand = "0.8.4"
 rand_distr = "0.4.1"
